added commit

+++ src/main/java/ru/netology/stats/StatsService.java
@@ -2,21 +2,21 @@ package ru.netology.stats;
 
 public class StatsService {
 
-    public int getSumSales(int[] sales) {
+    public int getSumSales(long[] sales) {
         int sumSales = 0;
-        for (int i : sales) {
+        for (long i : sales) {
             sumSales += i;
         }
         return sumSales;
     }
 
-    public int getAverageSumSales(int[] sales) {
+    public int getAverageSumSales(long[] sales) {
         int sumSales = getSumSales(sales);
         int averageSumSales = sumSales / 12;
         return averageSumSales;
     }
 
-    public int getMaxMonthSales(int[] sales) {
+    public int getMaxMonthSales(long[] sales) {
         int maxMonth = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] >= sales[maxMonth]) {
@@ -26,7 +26,7 @@ public class StatsService {
         return maxMonth + 1;
     }
 
-    public int getMinMonthSales(int[] sales) {
+    public int getMinMonthSales(long[] sales) {
         int minMonth = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] <= sales[minMonth]) {
@@ -36,7 +36,7 @@ public class StatsService {
         return minMonth + 1;
     }
 
-    public int getSumMonthBelowAverage(int[] sales) {
+    public int getSumMonthBelowAverage(long[] sales) {
         int averageSumSales = getAverageSumSales(sales);
         int sumMonthBelAv = 0;
         //int sumMonthAboveAverage = 0;
@@ -49,7 +49,7 @@ public class StatsService {
     }
 
 
-    public int getSumMonthAboveAverage(int[] sales) {
+    public int getSumMonthAboveAverage(long[] sales) {
         int averageSumSales = getAverageSumSales(sales);
         int sumMonthAbAv = 0;
         //int sumMonthAboveAverage = 0;
