+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -9,7 +9,7 @@ public class StatsServiceTest {
 
     public void shouldFindSumSales() {
         StatsService service = new StatsService();
-        int[] sales = {9, 36, 2, 16, 1, 20, 19, 8, 19, 22, 41, 15};
+        long[] sales = {9, 36, 2, 16, 1, 20, 19, 8, 19, 22, 41, 15};
         int expectedSumSales = 208;
         int actualSumSales = service.getSumSales(sales);
         Assertions.assertEquals(expectedSumSales, actualSumSales);
@@ -19,7 +19,7 @@ public class StatsServiceTest {
 
     public void shouldFindAverageSumSales() {
         StatsService service = new StatsService();
-        int[] sales = {9, 36, 2, 16, 1, 20, 19, 8, 19, 22, 41, 15};
+        long[] sales = {9, 36, 2, 16, 1, 20, 19, 8, 19, 22, 41, 15};
         int expectedAverageSumSales = 17;
         int actualAverageSumSales = service.getAverageSumSales(sales);
         Assertions.assertEquals(expectedAverageSumSales, actualAverageSumSales);
@@ -29,7 +29,7 @@ public class StatsServiceTest {
 
     public void shouldFindMaxMonthSales() {
         StatsService service = new StatsService();
-        int[] sales = {9, 36, 2, 16, 1, 20, 19, 8, 19, 22, 41, 15};
+        long[] sales = {9, 36, 2, 16, 1, 20, 19, 8, 19, 22, 41, 15};
         int expectedMaxMonthSales = 11;
         int actualMaxMonthSales = service.getMaxMonthSales(sales);
         Assertions.assertEquals(expectedMaxMonthSales, actualMaxMonthSales);
@@ -39,7 +39,7 @@ public class StatsServiceTest {
 
     public void shouldFindMinMonthSales() {
         StatsService service = new StatsService();
-        int[] sales = {9, 36, 2, 16, 1, 20, 19, 8, 19, 22, 41, 15};
+        long[] sales = {9, 36, 2, 16, 1, 20, 19, 8, 19, 22, 41, 15};
         int expectedMinMonthSales = 5;
         int actualMinMonthSales = service.getMinMonthSales(sales);
         Assertions.assertEquals(expectedMinMonthSales, actualMinMonthSales);
@@ -49,7 +49,7 @@ public class StatsServiceTest {
 
     public void shouldFindSumMonthBelowAverage() {
         StatsService service = new StatsService();
-        int[] sales = {9, 36, 2, 16, 1, 20, 19, 8, 19, 22, 41, 15};
+        long[] sales = {9, 36, 2, 16, 1, 20, 19, 8, 19, 22, 41, 15};
         int expectedSumMonthBelAv = 6;
         int actualSumMonthBelAv = service.getSumMonthBelowAverage(sales);
         Assertions.assertEquals(expectedSumMonthBelAv, actualSumMonthBelAv);
@@ -59,7 +59,7 @@ public class StatsServiceTest {
 
     public void shouldFindSumMonthAboveAverage() {
         StatsService service = new StatsService();
-        int[] sales = {9, 36, 2, 16, 1, 20, 19, 8, 19, 22, 41, 15};
+        long[] sales = {9, 36, 2, 16, 1, 20, 19, 8, 19, 22, 41, 15};
         int expectedSumMonthAbAv = 6;
         int actualSumMonthAbAv = service.getSumMonthAboveAverage(sales);
         Assertions.assertEquals(expectedSumMonthAbAv, actualSumMonthAbAv);
